refactor(angular): move router outlet to common
