+++ packages/angular/common/src/index.ts
@@ -20,3 +20,5 @@ export { bindLifecycleEvents, AngularDelegate } from './providers/angular-delega
 export type { IonicWindow } from './types/interfaces';
 
 export { NavParams } from './directives/navigation/nav-params';
+export { IonRouterOutlet, INPUT_BINDER, RoutedComponentInputBinder } from './directives/navigation/router-outlet';
+export type { StackEvent } from './directives/navigation/stack-utils';
